refactor(toolkit): split notices code into separate files
