+++ packages/@aws-cdk/toolkit-lib/lib/api/notices/index.ts
@@ -1,0 +1,1 @@
+export * from './notices';
